Move QuicContext to its own file to get it away from ifdefs
