+++ source/common/quic/quic_stat_names.h
@@ -2,7 +2,6 @@
 
 #ifdef ENVOY_ENABLE_QUIC
 
-#include "envoy/common/optref.h"
 #include "envoy/stats/scope.h"
 
 #include "source/common/common/thread.h"
@@ -12,11 +11,6 @@
 #include "quiche/quic/core/quic_types.h"
 
 namespace Envoy {
-
-namespace Server {
-class HotRestart;
-} // namespace Server
-
 namespace Quic {
 
 class QuicStatNames {
@@ -66,18 +60,6 @@ public:
 };
 
 #endif
-
-class QuicContext {
-public:
-  QuicContext(Stats::SymbolTable& stat_names, OptRef<Envoy::Server::HotRestart> hot_restart)
-      : stat_names_(stat_names), hot_restart_(hot_restart) {}
-  QuicStatNames& statNames() { return stat_names_; }
-  OptRef<Envoy::Server::HotRestart> hotRestart() { return hot_restart_; }
-
-private:
-  QuicStatNames stat_names_;
-  OptRef<Envoy::Server::HotRestart> hot_restart_;
-};
 
 } // namespace Quic
 } // namespace Envoy
