+++ source/server/server.h
@@ -34,7 +34,7 @@
 #include "source/common/init/manager_impl.h"
 #include "source/common/memory/heap_shrinker.h"
 #include "source/common/protobuf/message_validator_impl.h"
-#include "source/common/quic/quic_stat_names.h"
+#include "source/common/quic/quic_context.h"
 #include "source/common/router/context_impl.h"
 #include "source/common/runtime/runtime_impl.h"
 #include "source/common/secret/secret_manager_impl.h"
